feat: add onChildDraw to customize View's respond to user interaction & change behavior on click item in recycler view

+++ app/src/main/java/com/poc/android/myhospitals/main/MainActivityFragment.java
@@ -3,6 +3,9 @@ package com.poc.android.myhospitals.main;
 import android.arch.lifecycle.Observer;
 import android.content.Intent;
 import android.graphics.Canvas;
+import android.graphics.Color;
+import android.graphics.Paint;
+import android.net.Uri;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
 import android.os.Bundle;
@@ -27,15 +30,17 @@ public class MainActivityFragment extends Fragment implements ItemListAdapter.Li
 
     private ItemViewModel mItemViewModel;
     public static final int EDIT_ACTIVITY_REQUEST_CODE = 2;
+    Bundle mBundle;
 
     // Bundle Key
     public static final String ITEM_INFO = "3";
+    public static final String ITEM_URL = "2";
 
     public MainActivityFragment() {
     }
 
     @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+    public View onCreateView(LayoutInflater inflater, final ViewGroup container,
                              Bundle savedInstanceState) {
         View rootView = inflater.inflate(R.layout.fragment_main, container, false);
 
@@ -70,11 +75,45 @@ public class MainActivityFragment extends Fragment implements ItemListAdapter.Li
 
                     @Override
                     public void onSwiped(RecyclerView.ViewHolder viewHolder, int direction) {
-                        final int fromPos = viewHolder.getAdapterPosition();
-                        Item myItem = adapter.getItemAtPosition(fromPos);
 
-                        // delete the item
-                        mItemViewModel.deleteItem(myItem);
+                        // RIGHT to LEFT :: delete item
+                        if (direction == ItemTouchHelper.LEFT) {
+                            final int fromPos = viewHolder.getAdapterPosition();
+                            Item myItem = adapter.getItemAtPosition(fromPos);
+
+                            // delete the item
+                            mItemViewModel.deleteItem(myItem);
+                        } else { // LEFT to RIGHT :: edit item
+                            Intent intent = new Intent(getActivity(), NewItemActivity.class);
+                            intent.putExtra(ITEM_INFO, mBundle);
+                            getActivity().startActivityForResult(intent, EDIT_ACTIVITY_REQUEST_CODE);
+                        }
+                    }
+
+                    @Override
+                    public void onChildDraw(Canvas c, RecyclerView recyclerView, RecyclerView.ViewHolder viewHolder, float dX, float dY, int actionState, boolean isCurrentlyActive) {
+
+                        if (actionState == ItemTouchHelper.ACTION_STATE_SWIPE) {
+                            // Get RecyclerView item from the ViewHolder
+                            View itemView = viewHolder.itemView;
+                            Paint p = new Paint();
+                            if (dX > 0) {
+                                /* Set color for positive displacement */
+                                p.setColor(Color.GREEN);
+
+                                // Draw Rect with varying right side, equal to displacement dX
+                                c.drawRect((float) itemView.getLeft(), (float) itemView.getTop(), dX,
+                                        (float) itemView.getBottom(), p);
+                            } else {
+                                /* Set color for negative displacement */
+                                p.setColor(Color.RED);
+
+                                // Draw Rect with varying left side, equal to the item's right side plus negative displacement dX
+                                c.drawRect((float) itemView.getRight() + dX, (float) itemView.getTop(),
+                                        (float) itemView.getRight(), (float) itemView.getBottom(), p);
+                            }
+                            super.onChildDraw(c, recyclerView, viewHolder, dX, dY, actionState, isCurrentlyActive);
+                        }
                     }
                 });
         helper.attachToRecyclerView(recyclerView);
@@ -83,13 +122,11 @@ public class MainActivityFragment extends Fragment implements ItemListAdapter.Li
 
     @Override
     public void onItemClick(Bundle bundle) {
-//        Uri webPage = Uri.parse(url);
-//        Intent intent = new Intent(Intent.ACTION_VIEW, webPage);
-//        if (intent.resolveActivity(getActivity().getPackageManager()) != null) {
-//            startActivity(intent);
-//        }
-        Intent intent = new Intent(getActivity(), NewItemActivity.class);
-        intent.putExtra(ITEM_INFO, bundle);
-        getActivity().startActivityForResult(intent, EDIT_ACTIVITY_REQUEST_CODE);
+        mBundle = bundle;
+        Uri webPage = Uri.parse(bundle.getString(ITEM_URL));
+        Intent intent = new Intent(Intent.ACTION_VIEW, webPage);
+        if (intent.resolveActivity(getActivity().getPackageManager()) != null) {
+            startActivity(intent);
+        }
     }
 }
